Leetcode problem 454
sum of A and B is subtracted from zero. Then, it is kept in dictionary. If sum matches dictionary key, the count is incremented by 1. Then, sum of C and D is searched in dictionary. If sum of C and D matches with any dictionary key, total count of tuples is incremented using value of the dictionary key.